+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * Remove the depth limit for null move pruning (5.26 +- 3.83)
+* Stop using Result to indicate stopped searches (1.80 +- 3.53)
 
 ## [8.0]